ui: pool members: add filter for name, node or vmid when adding guests
Makes it easier to find specific guests when one got many of them.

I decided to not allow filtering the current pool name, as often there
is none set for the guest the user wants to add and one could
theoretically sort the current pool column to group guests by them,
and finally, we can still add that easily if it gets requested with a
good enough rationale.

Signed-off-by: Thomas Lamprecht <t.lamprecht@proxmox.com>