fix #6779: pveupdate: renew already expired certificates
if nodes are offline for a longer period of time, they might not be renewed by
pveupdate before they expire. the `verify` call here just serves as an
extra safeguard to prevent accidental overwriting of certificates not actually
signed by the cluster CA, checking the expiry time servers no purpose.

Suggested-by: Stephane Chazelas
Reviewed-by: Fiona Ebner <f.ebner@proxmox.com>
Link: https://lore.proxmox.com/all/20250909100529.415836-1-f.gruenbichler@proxmox.com
Signed-off-by: Fabian Grünbichler <f.gruenbichler@proxmox.com>