api: node shells: make helper aware of tunnel command
Rename the not so telling $remcmd to $tunnel_cmd and make the helper
prefix it to the actual command.

This is a preparation of adapting login for proxied (tunneled) shells,
i.e. where the user requests a shell from another node than they
opened the web UI for.

No semantic change intended.

Signed-off-by: Thomas Lamprecht <t.lamprecht@proxmox.com>