api: node shells: avoid nested login for already logged in root@pam
When using the vncshell or xtermjs based shells we use a SSH tunnel if
the requested node is different from the one the request was made too.

Further we special case sessions from the root@pam user, as they
already got verified, so do not need to re-login on the target shell.

With Debian Trixie based releases we had to make a change to make this
shell work again [0], it seems that there is still a race in the
whole interaction, at least if there is a nested login shell. See [1]
for more details.

The Debian Trixie version of the `login` manual page explicitly
mentions a bug related to this:

> A recursive login, as used to be possible in the good old days, no
> longer works; for most purposes su(1) is a satisfactory substitute.
> Indeed, for security reasons, login does a vhangup(2) system call
> to remove any possible listening processes on the tty. This is to
> avoid password sniffing. If one uses the command login, then the
> surrounding shell gets killed by vhangup(2) because it’s no longer
> the true owner of the tty. This can be avoided by using exec login
> in a top-level shell or xterm.
-- man 1 login

IIRC this was checked back when implanting [0], but as it was during
quite eventful bootstrapping times of Debian Trixie releases I'm not
100% certain about that.

If the issues our users sometimes (?) see stem indeed from a race
related to above bug, we should be able to avoid that by dropping the
explicit login call for root@pam when tunneling. Other @pam users
would be still affected, but as the partial fix is so simple and
correct in any way it's still worth roll it out sooner.

To improve this more broadly the following two options seem most
promising:

1. Replace the manual SSH tunnel by our proxy-to-node infrastructure
   and tunnel the websocket of the target nodes termproxy command
   through that.

2. Use a wrapper tool to handle the login command such that it does
   not interferes with the outer login, could potentially even be an
   existing one like dtach, or alternatively something we write
   ourselves.

I did not yet evaluate impact and work required for either option to
happen, but from an experienced gut feeling the first option would be
the better one, especially as we want to drop SSH for tunneling
completely in the mid to long term.

As this is not a complete and certain fix for #6789 [1] and especially
as we cannot really reproduce this ourselves here in any useful way,
I refrained from adding a fix # to the commit message, but it should
be a partial fix.

[0]: https://git.proxmox.com/?p=pve-xtermjs.git;a=commitdiff;h=7b47cca8368e63c30f6227442570f9f35dd7ccf0
[1]: https://bugzilla.proxmox.com/show_bug.cgi?id=6789

Signed-off-by: Thomas Lamprecht <t.lamprecht@proxmox.com>