fix #2077: pvenode: remove dependency on libterm-readline-gnu-perl
Switch from the external Term::ReadLine to our PVE::PTY module.

For one this avoids an extra dependency when we got our own method
that – while having a far smaller feature set – is very robust and
does the basic job of prompting on a PTY well.

But with Debian Trixie and PVE 9 we got actually another more pressing
reason: Term::ReadLine had always problems if more than one instance
was active at the same time, but with the version from trixie they now
actually check for that and throw an exception on instantiation if the
library detects that it was already instantiated once.

We had such a nested usage when using pvenode CLI to order a TLS cert
through ACME that was (seemingly) working OK before, but got broken by
this new assertion, so while #2077 wanted just to drop the dependency
for better code reuse, this now actually fixes #6748 too,

[0]: https://github.com/hirooih/perl-trg/pull/20

Originally-by: Stefan Hrdlicka <s.hrdlicka@proxmox.com>
[FE: rebase and adapt to rename]
Signed-off-by: Fiona Ebner <f.ebner@proxmox.com>
Link: https://lore.proxmox.com/20250918135215.95188-4-f.ebner@proxmox.com
 [TL: add commit message with context]
Signed-off-by: Thomas Lamprecht <t.lamprecht@proxmox.com>