ui: ha: remove group info from HA status string in guest status view
Groups have been deprecated in favor of rules, so this information
should no longer be displayed. Since there is no longer a 1:1 mapping of
resources to groups and resources could be in many rules, displaying
information about the rules in this view would clutter the guest status
view. In addition, finding the rules of which the resource is a part of
adds a lot more complexity than getting the group did. Thus, the
information is omitted from the status text altogether.

Signed-off-by: Michael Köppl <m.koeppl@proxmox.com>
Link: https://lore.proxmox.com/20250919140620.243497-2-m.koeppl@proxmox.com