fix #6815: ui: resource tree: update tag colors properly
In commit:
 e19de81f (ui: resource tree: fix change detection)
the change detection for resource tree nodes was fixed so that it did
not update on every refresh anymore.

This now also means that when the ui options are loaded (which contains
the tag color overrides) we now have to correctly update nodes which
contains tags, otherwise the standard colors will be shown.

There is already an event handler for the 'loadedUiOptions', so simply
adapt it to also update the tree nodes that contains tags.

Fixes: e19de81f (ui: resource tree: fix change detection)
Signed-off-by: Dominik Csapak <d.csapak@proxmox.com>
Link: https://lore.proxmox.com/20251010072802.798062-1-d.csapak@proxmox.com