ui: add context to translatable strings
These will be displayed in translator editors and provide context for
translators.

Some strings like "Verify State" are not clear in meaning without added
context. It is a reasonable assumption for "Verify" to be a verb, since
button labels should start with a verb and we do not have enough context
in the i18n context to see this is not part of such a label, but in this
case we actually mean the state of the verification task.

Signed-off-by: Maximiliano Sandoval <m.sandoval@proxmox.com>
Reviewed-by: Shannon Sterz <s.sterz@proxmox.com>
Link: https://lore.kernel.org/r/20251023083054.116763-1-m.sandoval@proxmox.com
[FE: add 'ui' prefix to commit title
     slightly improve wording for otel-collector and volume-chain
     make tidy]
Signed-off-by: Fiona Ebner <f.ebner@proxmox.com>