ui: properly reset API 401 count
due to a timing race condition, it can happen that we try to increase
the 401 count before resetting it (which happens after the first
ticket api call). Since `undefined + ` === NaN, our later check for
`> 5` never triggers and we don't show the login mask, but an empty
window.

To fix this, initialize the variable always with 0.

Observed when restoring a session after a ticket expired while the
browser was closed.

Signed-off-by: Dominik Csapak <d.csapak@proxmox.com>
Reviewed-by: Fiona Ebner <f.ebner@proxmox.com>
Link: https://lore.proxmox.com/20251029131256.3124803-1-d.csapak@proxmox.com