ui: resource tree: prevent overwriting of 'text' property
currently, when 'setText' is called for a tree item, the 'text' property
in the store is overwritten. This can occur when the item changes or
when the 'UIOptions' change, which can be triggered arbitrarily (e.g. by
opening the 'Datacenter -> Options' panel.)

Since this can be triggered manually, overwriting the 'text' property
can lead to tags being shown multiple times. (When sorting is set to
vmid).

To prevent this from happening, instead of overwriting the 'text'
property, use the 'renderer' method of the column. For that the column
has to be manually redefined.

This should prevent similar problems in the future, since the 'text' is
not overwritten anymore, but it is generated during rendering.

Fixes: 952a4f61 (fix #6815: ui: resource tree: update tag colors properly)
Reported-by: Stefan Hanreich <s.hanreich@proxmox.com>
Signed-off-by: Dominik Csapak <d.csapak@proxmox.com>
Link: https://lore.proxmox.com/20251107140136.3972272-1-d.csapak@proxmox.com
(cherry picked from commit 83783c3b1e27e4750d475cd62342154f19b445c5)
Signed-off-by: Thomas Lamprecht <t.lamprecht@proxmox.com>