ui: cpu flag selector: query CPU flag list via API
There is coupling between the store and value for the top-level grid,
which has a field mixin. Because of this, the setValue() method needs
to update the 'state' value for the records in the store accordingly.
This can however only be done if the store is already loaded, so the
store gets a refresh() callback. It also remembers which value it was
already adjusted for for potential subsequent setValue() calls.

Signed-off-by: Fiona Ebner <f.ebner@proxmox.com>
Link: https://lore.proxmox.com/20251107144429.126790-9-f.ebner@proxmox.com