refactor: ui: sdn browser: parametrize zone content panel
In order to reuse the zone content panel for future, additional,
panels in the SDN browser, move all static configuration properties
into the declaration of the ExtJS widget. In particular, make the
on_(de)select callbacks configurable, so this panel can be customized
to work with a multitude of additional panels.

This is a preparation for extended status reporting for SDN zones,
which will add several new panels that want to utilize the content
view.

The selectionModel uses the default configuration that is included in
the GridPanel, so it can be removed.

No functional changes intended.

Signed-off-by: Stefan Hanreich <s.hanreich@proxmox.com>
Link: https://lore.proxmox.com/20251113161926.793020-14-s.hanreich@proxmox.com