pvestatd: add network resource to status reporting
The new network resource will act as the top-level resource for all
networking entities (including SDN entities). The network resource
contains a network_type field, which indicates the type of networking
resource - similar to how the storage plugin handles different types
of storages. For now, it contains SDN fabrics and the SDN zones have
been copied over as well.

The main reason for moving over to a new resource type is the current
ID schema of the SDN resource, which is 'sdn/{zone_id}'. This makes it
hard to extend without the possibility of ID collisions. Additionally,
since the ID is used in several places throughout the backend / UI,
changing the schema would break compatibility with nodes that are on
an earlier version and would be an API break as well.

With this patch, nodes will still broadcast the old format for
backwards-compatibility. A subsequent patch for the resources API will
handle this case and ignore the old SDN resources from nodes that are
already sending the new format, to avoid including zones twice in the
returned resource data.

Co-authored-by: Stefan Hanreich <s.hanreich@proxmox.com>
Signed-off-by: Gabriel Goller <g.goller@proxmox.com>
Signed-off-by: Stefan Hanreich <s.hanreich@proxmox.com>
Link: https://lore.proxmox.com/20251113161926.793020-15-s.hanreich@proxmox.com