api: cluster: add support for network resource type
pvestatd now broadcasts a new network resource type, instead of the
sdn resource type. This commit adds handling for this new network type
to the resources endpoint. In order to be able to deal with older
nodes, keep support for parsing the old sdn resource type.

Upgraded nodes will still broadcast the old format for
backwards-compatibility and nodes with this patch applied support
handling both formats. With this patch, nodes will check whether a
node is sending both formats or only the old one, and parse the
resources based on that information. Older nodes will drop the new
network resource type, but will still be able to show zones, because
the old format still gets broadcast. Newer nodes will take the
information from the network store, if available, otherwise fall back
to the SDN store.

Another reason for keeping the old format around is so we do not break
older clients, that rely on the old SDN format - removing it would be
a breaking API change.

Signed-off-by: Stefan Hanreich <s.hanreich@proxmox.com>
Link: https://lore.proxmox.com/20251113161926.793020-16-s.hanreich@proxmox.com