pvestatd: sdn: adapt to changes in status reporting
Since pve-network now includes the fabric status in its top-level
status function call (SDN::status), move pvestatd over to use the
zone-specific function for generating SDN resources instead.
Additionally, the zone_type field has been added in the output of the
zone status, so broadcast and return that field as well. Main use-case
for now is discerning the type of zones, to decide which panels to
show in the SDN browser.

Also, the have_sdn guard can be removed since SDN is pre-installed and
therefore the guard is no longer needed.

Signed-off-by: Stefan Hanreich <s.hanreich@proxmox.com>
Link: https://lore.proxmox.com/20251113161926.793020-17-s.hanreich@proxmox.com