ui: resource tree: add network resource
Add the newly introduced network resource to the resource tree, so the
the status of network resources can be displayed in the UI. For this
matter, a new NetworkBrowser widget is added, which is responsible for
showing the contents of a network resource. The old SDN browser is
still kept around, so the UI can handle nodes that are still sending
the old sdn resource type.

Co-authored-by: Stefan Hanreich <s.hanreich@proxmox.com>
Signed-off-by: Gabriel Goller <g.goller@proxmox.com>
Signed-off-by: Stefan Hanreich <s.hanreich@proxmox.com>
Link: https://lore.proxmox.com/20251113161926.793020-18-s.hanreich@proxmox.com