ui: network browser: add zone bridge view
This shows a list of all bridges, that are part of an SDN zone
(including the pseudo localnetwork zone, that corresponds to the local
network of a node). It also shows additional information about which
ports are currently members of the bridge, as well as their VLAN
configuration and which guest they belong to.

Signed-off-by: Stefan Hanreich <s.hanreich@proxmox.com>
Link: https://lore.proxmox.com/20251113161926.793020-21-s.hanreich@proxmox.com