api: metric export: re-use credentials from request for proxying
This avoids the need to proxy to the privileged API daemon to assemble
a ticket for the user manually.
Re-using the exact same credentials as the original request also
ensures that the permission set stays the same and fully correct for
API tokens, while this wasn't a problem in practice here, it's still
much better to play safe - especially if this is copied/moved over
elsewhere for another use case where it then could have problematic
implications.

The performance gain i expected due to proxying less sadly did not
materialize in my tests with large amount of metrics (e.g., due to
having 10k guests).

Signed-off-by: Thomas Lamprecht <t.lamprecht@proxmox.com>