guest: refactor and reuse AgentIPView for containers
Create a common base for the IP view and two sub-classes, one for QEMU
VMs and for LXC CTs.

In the generic guest summary widget we can now simply switch the xtype
of the IP widget between the LXC or QEMU one depending on the guest
type.

With that, the container Summary page now also shows the IP addresses
that the container has configured.

Also increase the (min) height of the guest summary cards (status,
notes and RRD metric graphs) from 330px to 360px to ensure that there
is enough vertical space for the IPs and the "More" button in the
status card.

With the increase in height the charts have now 250px height available
for the actual graphing canvas, which makes them look slightly nicer
and clearer, especially given the recent increase of RRD granularity.

Co-authored-by: Gabriel Goller <g.goller@proxmox.com>
Link: https://lore.proxmox.com/20250527140446.365983-3-g.goller@proxmox.com
 [TL: rework approach to favor specialisation over callbacks]
Signed-off-by: Thomas Lamprecht <t.lamprecht@proxmox.com>