ui: ip view: ensure container is fully initialized before store update
It seemingly can happen that this is called before the container is
rendered, and then the query method throws tries to deference some
items.items.length somewhere in the stack that is not defined, causing
the whole UI to break.

It's rather odd and I cannot really phantom why ExtJS breaks down so
badly here in that case, as it seems rather brittle to assume that all
components that are searched on query are fully instantiated, maybe
there something else broken too; but this is at least a stop gap..

I only ran into this for QEMU, but that might have just been (bad)
luck.

Signed-off-by: Thomas Lamprecht <t.lamprecht@proxmox.com>