ui: lxc: add support to edit container runtime environment
With a recent pve-container we can now configure the environment
(variables) of a container through a new `env` property, which is a
NUL-separated list of key=value options that make up the separate
environment variables.

Add initial support to configure these through the web UI. This is
done through adding a simple component based of a field container that
makes up a single environment variable and using that in an input
panel. There we simply split the NUL-separated list and create a
EnvVar field of each entry.
Allow adding new env-variable entries and allow deleting them.
For now disable form reset tracking, that needs a bit better state
handling internally, for which I simply did not have the time to look
into more closely, besides tracking original values there is also the
question of how to "revive" deleted fields, the simplest option might
be to actually handle this centrally in the input panel by just
recreating the original state there. Anyhow, for this component it's
not really that problematic to have no reset form, in the worst case a
user need to close and re-open the dialog, not great but given the
time I got left for more important things it's an acceptable
workaround.

Signed-off-by: Thomas Lamprecht <t.lamprecht@proxmox.com>