ui: lxc network: support advanced host-managed option
The new host-managed flag is mainly intended for CT created from OCI
images that do not have their own network management/config stack.

pve-container will auto-enable it if the init cmd points to something
else than /sbin/init as heuristic, but with exposing it here a user
can force it on too.

Signed-off-by: Thomas Lamprecht <t.lamprecht@proxmox.com>