ui: add context menu with bulk actions for datactenter in resource tree
Now that we have guest bulk-actions for the whole cluster, independent
of where the guests reside, we can provide a context menu for the
top-level Datacenter root node of the resource tree, providing a
convenient way to access this functionality from basically everywhere.

Copied and adapted from the PVE.node.CmdMenu node one. We probably can
provide a common base here, but these things are rather static anyway,
so this won't matter _that_ much.

Signed-off-by: Thomas Lamprecht <t.lamprecht@proxmox.com>