ui: add context menu with bulk actions for tags in resource tree
The resource tree provides a tag view, with which all guests are
grouped by their tags. With the new cluster-wide bulk-actions we can
relatively easily provide a context menu for the tag tree-node itself.

The new provides the all guest bulk-actions, just like the recently
added datacenter context menu, but it will open the bulk actions with
the "include tags" filter pre-selected for the tag one opened the
context menu on.

The implementation mostly reuses the datacenter menu, which gets
slightly adapted to support that.

Signed-off-by: Thomas Lamprecht <t.lamprecht@proxmox.com>