ui: bulk actions: don't try to reset non-existing 'vmidfilter'
there never was a field with the id 'vmidfilter', so it was either a
left over from a previous attempt, or a copy&paste mistake.

It does not fix any behavior, since it was the last thing to be done in
the function anyway, but it prevents an exception/error log in the console.

Signed-off-by: Dominik Csapak <d.csapak@proxmox.com>
Link: https://lore.kernel.org/r/20251117100336.894345-1-d.csapak@proxmox.com