Add support for TDX attestation
Add configuration options for the quote-generation-socket used to communicate
with the Quote Generation Service. This enables attestation support for TDX
using vsock.

More information can be found at:
https://www.qemu.org/docs/master/interop/qemu-storage-daemon-qmp-ref.html#object-QSD-qom.TdxGuestProperties

This only includes support for vsock. The SocketAddress object in QEMU also
supports unix, inet and fd, which are not commonly used in TDX.

Signed-off-by: Anton Iacobaeus <anton.iacobaeus@canarybit.eu>
Link: https://lore.proxmox.com/20251117104817.471815-2-anton.iacobaeus@canarybit.eu
[FE: do not expose socket type]
Signed-off-by: Fiona Ebner <f.ebner@proxmox.com>