fix: ui: trim OCI registry reference
Trim the input such that an added space (e.g. from copying the URL from
a registry page) does not yield error "reference: value does not match
the regex pattern". Although true that the regex is then not matched,
this change avoids failing the validation due to an additional symbol
that can simply be trimmed.

Signed-off-by: Michael Köppl <m.koeppl@proxmox.com>
Link: https://lore.proxmox.com/20251117124553.138750-1-m.koeppl@proxmox.com