ui: resource tree: fix rendering of guest text when sorted by name
In a recent commit, we changed how we render the text for element in the
resource tree, namely by not overwriting the original data in the store.

During that code move & change, a bug slipped in, where we correctly
detected that we're sorting by name and set the text to the correct
format, but ultimately discarded that value and overwrote it with the
original one from the store.

Fix this by first extracting the original text from the record, and the
use the text like we did the record field before

This was reported in the forum:
https://forum.proxmox.com/threads/176021/

Fixes: 83783c3b (ui: resource tree: prevent overwriting of 'text' property)
Signed-off-by: Dominik Csapak <d.csapak@proxmox.com>
Link: https://lore.proxmox.com/20251117131945.2148025-1-d.csapak@proxmox.com
Signed-off-by: Thomas Lamprecht <t.lamprecht@proxmox.com>