ui: resource tree: render clustername consistently
since we changed the tree to create the store outside the component on
first load, setting the clustername on the root element of the tree
could be race against the recreation of that node on the first load.

This made the display of the cluster name inconsistent, iow. sometimes
the clustername vanished after the first load.

To prevent that, don't set the clustername on the store record directly,
but adapt the render method to add the clustername from the global
`PVE.ClusterName` variable if it exists. Since this happens on login
(which happens before the load for the resource tree is triggered)
it happens before the rendering.

Reported-by: Thomas Lamprecht <t.lamprecht@proxmox.com>
Signed-off-by: Dominik Csapak <d.csapak@proxmox.com>
Link: https://lore.proxmox.com/20251117153721.3161092-1-d.csapak@proxmox.com
Signed-off-by: Thomas Lamprecht <t.lamprecht@proxmox.com>
(cherry picked from commit 5d345fab0b48d9e9da820cb45c98bf8a3dc71b68)
Signed-off-by: Thomas Lamprecht <t.lamprecht@proxmox.com>