ui: create wizard: only submit ha-managed if checked for better compat
Avoid sending the parameter with the default value when it's not
necessary, improving compatibility in cases where the ha-managed param
is not recognized (e.g. when creating a guest on another node that does
not support the parameter). Setting uncheckedValue to undefined results
in the value of the checkbox not being submitted upon submission of the
form if the checkbox is left unchecked or explicitly set unchecked [0].
The API already considers the ha-managed param optional and sets a
default value of 0.

Even though undefined is the default value for uncheckedValue, set it
explicitly to make the intent clearer.

[0] https://docs.sencha.com/extjs/7.0.0/classic/Ext.form.field.Checkbox.html#cfg-uncheckedValue

Signed-off-by: Michael Köppl <m.koeppl@proxmox.com>
Link: https://lore.proxmox.com/20251117173321.240894-1-m.koeppl@proxmox.com