ui: console: detect and warn on application containers
application containers often don't have a functional shell (since it
runs directly, without init) so try to detect and warn the users
in the inline console.

The detection heuristic is:
* cmode is 'console'
* entrypiont is not '/sbin/init'

Those values are not the (lxc) default and should cover most application
containers.

Since we change the layout a bit, focus the actual iframe component, not
the panel.

Signed-off-by: Dominik Csapak <d.csapak@proxmox.com>
Link: https://lore.proxmox.com/20251118130710.2362914-1-d.csapak@proxmox.com
Signed-off-by: Thomas Lamprecht <t.lamprecht@proxmox.com>