Revert "fix #6815: ui: resource tree: update tag colors properly"
This reverts commit 952a4f615731c3de0351c94b83bd97340c7e7798.

Since we moved the text rendering to a renderer instead of overwriting
the 'text' property on adding the elements to the tree, this fix is
not necessary anymore since the renderer has the correct information
already.

This also wrongly rendered the guest entries in the tag view with their
'groupRenderer' instead of their normal text (vmid + name).

Signed-off-by: Dominik Csapak <d.csapak@proxmox.com>
Link: https://lore.proxmox.com/20251118133517.2540651-1-d.csapak@proxmox.com
Signed-off-by: Thomas Lamprecht <t.lamprecht@proxmox.com>