ui: storage oci pull: reword tag placeholder to not suggest its the default
Before we just used "latest" as empty text, and it looked like this
was an implied default value, which made the validity error due to
this field being left blank a bit confusing.

While we potentially might even just use that as default, seems most
OCI images on registries do provide an image under that tag, that's
rather a decision for another day, making the user chose is
definitively a bit safer too.

Reported-by: Dominik Csapak <d.csapak@proxmox.com>
Signed-off-by: Thomas Lamprecht <t.lamprecht@proxmox.com>