ui: guest console: hide the application container hint after 8s
Some OCI images that are for full system images set /bin/bash as their
entrypoint, and for these the console works (rudimentary, as the CT
shuts down if one fully exits the bash instance). And alarming hints
that cannot be dismissed are rather bad UX, as simple stop gap we can
at least hide it after a medium-short time. Chose 8s as that feels
long enough to have enough time to read it twice but not that long to
make it overly annoying.

Signed-off-by: Thomas Lamprecht <t.lamprecht@proxmox.com>