api: cluster bulk actions: increase default worker to 4 for start, stop & suspend
As otherwise this just way to slow for any real workload.

Partially fixes #7048, but needs to have smarter selection or
scheduling to fully address that enhancement request.

Do not increase the default for migrations, as that can induce network
traffic on top of storage and CPU load, and as we always have a single
target it's not something we want to have a higher default for.
Migrate is also the only bulk action that exposes the max-worker
setting in the frontend.

Signed-off-by: Thomas Lamprecht <t.lamprecht@proxmox.com>