pveupdate: do not try to update subscription when using offline key
pveupdate is run daily by pve-daily-update.service and, among other
things, it tries to update the subscription. Updating a subscription
fails early when an offline key is installed, hence, pveupdate will
currently log the following to the syslog with severity level "error":

> update subscription info failed: Updating offline key not possible -
> please remove and re-add subscription key to switch to online key.

This may be confusing to users, because it sounds like an action is
needed on their part to resolve an error.

To avoid that, skip the subscription update if an offline key is
installed.

Reported in enterprise support.

Signed-off-by: Friedrich Weber <f.weber@proxmox.com>
Link: https://lore.proxmox.com/20251127092617.60723-1-f.weber@proxmox.com
Signed-off-by: Thomas Lamprecht <t.lamprecht@proxmox.com>