api: node bulk action: improve cpu-thread-count derived max-worker fallback
Modern server systems can have _many_ CPU threads (like over 256),
which is rather high max-worker count, as that can cause lots of (IO
and CPU) pressure.

To make the fallback a bit more easy on the system deduct one from the
core-count to have some headroom for the system and clamp the maximum
auto-default to 8 for now, which is still relatively high.

Signed-off-by: Thomas Lamprecht <t.lamprecht@proxmox.com>