ui: bulk action window: add max-workers field to all actions
This was only added to the migrate actions, but we now have broad
support for all bulk action types for both cluster level and node
level, so let users leverage that.

Allow the field to be blank as with that we have better backward
compatibility if there are older nodes  in a cluster, which can easily
happen during an upgrading each node one per one and using the
bulk-actions to move or shutting down guests when doing so.
Also use a generic "auto" placeholder text, as the what is used as
actual value is a bit more complex due to multiple fallback levels and
some difference for these between cluster level and node level
actions.

Use same flex ratio as the bulk stop action to have the fields line
up there, for the other actions, where no other fields are present,
this still looks OK.

Signed-off-by: Thomas Lamprecht <t.lamprecht@proxmox.com>