ui: bulk action window: record todo comment to adapt max-worker field name
This was the only long-time pre-existing one, so do not use the new
one already. The reason here is that especially this action is often
used when doing cluster wide upgrades to free a node from all guests,
so if we would adapt to the new name already, migrations from any
older node to any node won't work if the user has loaded the web UI
from an already upgraded node.

Signed-off-by: Thomas Lamprecht <t.lamprecht@proxmox.com>